Move timing functions to tests
