+++ setup.py
@@ -14,5 +14,6 @@ setup(
         'google-api-python-client==1.6.6',
         'google-auth-httplib2==0.0.3',  # required for using google.auth credentials with google-api-python-client
         'requests==2.18.1',
+        'future==0.16.0',
     ],
 )
